Fix build error: Separate client/server components in Header
- Keep Header.tsx as server component to allow MenuWrapper import
- Create new HeaderContent.tsx client component for route-aware styling
- Fixes bcrypt/node-pre-gyp bundling issue in client components
